Show the attested origin on signing screens
Verification was shown only when connecting. Every signing prompt after that
displayed the dApp's self-reported url alone, which is the one value a
phishing site fully controls, while the attestation that contradicts it sat
unused on the request.

Share the alert and render it on all four request screens. They report
impersonation and known scams but not merely unverified origins: the origin
was accepted when the session was approved, and repeating that on every
signature would crowd out the two warnings that carry weight.
