+++ walletkit-chain-evm/src/main/java/io/horizontalsystems/walletkit/modules/walletconnect/request/sendtransaction/WCSendEthScreen.kt
@@ -41,6 +41,7 @@ import io.horizontalsystems.walletkit.ui.compose.components.VSpacer
 import io.horizontalsystems.walletkit.ui.compose.components.headline1_leah
 import io.horizontalsystems.walletkit.ui.compose.components.rememberAsyncAction
 import io.horizontalsystems.walletkit.ui.compose.components.subhead_grey
+import io.horizontalsystems.walletkit.modules.walletconnect.VerificationAlert
 import io.horizontalsystems.walletkit.ui.helpers.TextHelper
 import io.horizontalsystems.walletkit.uiv3.components.AlertCard
 import io.horizontalsystems.walletkit.uiv3.components.AlertFormat
@@ -136,6 +137,12 @@ fun WCSendEthRequestScreen(
                 modifier = Modifier.fillMaxWidth()
             )
             VSpacer(16.dp)
+
+            // The url above is what the dApp says it is. This is the attested origin, and it is the
+            // only thing on the screen a phishing site cannot set. Unverified origins are not
+            // reported here: that was already accepted when the session was approved.
+            VerificationAlert(sessionRequestUI.verification, reportUnverified = false)
+
             Column(
                 modifier = Modifier
                     .padding(vertical = 8.dp, horizontal = 16.dp)
