+++ walletkit/src/main/java/io/horizontalsystems/walletkit/modules/walletconnect/request/WcRequestScreen.kt
@@ -35,6 +35,7 @@ import io.horizontalsystems.walletkit.ui.compose.ComposeAppTheme
 import io.horizontalsystems.walletkit.ui.compose.components.VSpacer
 import io.horizontalsystems.walletkit.ui.compose.components.headline1_leah
 import io.horizontalsystems.walletkit.ui.compose.components.subhead_grey
+import io.horizontalsystems.walletkit.modules.walletconnect.VerificationAlert
 import io.horizontalsystems.walletkit.ui.helpers.TextHelper
 import io.horizontalsystems.walletkit.uiv3.components.bottombars.ButtonsGroupHorizontal
 import io.horizontalsystems.walletkit.uiv3.components.bottomsheet.BottomSheetContent
@@ -125,6 +126,11 @@ fun WcRequestScreen(
                     modifier = Modifier.fillMaxWidth()
                 )
             }
+
+            // The url above is what the dApp says it is. This is the attested origin, and it is the
+            // only thing on the screen a phishing site cannot set. Unverified origins are not
+            // reported here: that was already accepted when the session was approved.
+            VerificationAlert(sessionRequest.verification, reportUnverified = false)
             VSpacer(16.dp)
             Column(
                 modifier = Modifier
